ipc port finished for riscv

+++ inc/riscv/env.h
@@ -68,7 +68,7 @@ struct Env {
 	// Lab 4 IPC
 	bool env_ipc_recving;		// Env is blocked receiving
 	void *env_ipc_dstva;		// VA at which to map received page
-	uint32_t env_ipc_value;		// Data value sent to us
+	uint64_t env_ipc_value;		// Data value sent to us
 	envid_t env_ipc_from;		// envid of the sender
 	int env_ipc_perm;		// Perm of page mapping received
 };
